More code consolidation
